edit logic for added and delete review

+++ app/db/models.py
@@ -66,7 +66,7 @@ class Markets(Base):
     id = Column(Integer, primary_key=True)
     fmid = Column(Integer, unique=True)
     name = Column(String(100), nullable=False)
-    rating = Column(Integer)
+    rating = Column(Float)
     website = Column(Text)
     street = Column(Text)
     city_id = Column(Integer, ForeignKey("cities.id"))
